refactor core hotspot files
Closes #209
