+++ .gitignore
@@ -5,3 +5,4 @@ dist
 .DS_Store
 setup-repo.ts
 tsconfig.tsbuildinfo
+.fallow/
