+++ ci/fallow-health-baseline.json
@@ -1,0 +1,4 @@
+{
+  "healthScore": 60,
+  "hotspotCount": 32
+}
